capabilities: CapLimit is inherited and does not depend on CapBound 599a973919c296c9134c76635e53a00b74bc8dc1

+++ src/common.hpp
@@ -8,7 +8,7 @@
 
 #define noinline __attribute__((noinline))
 
-#define BIT(nr) (1ULL << (nr))
+#define BIT(nr) ((uint64_t)1ULL << (nr))
 
 #ifndef __has_feature
 # define __has_feature(__x) 0
